move attributes to utils
